+++ pandapower/test/__init__.py
@@ -7,4 +7,3 @@ tutorials_path = os.path.join(os.path.dirname(pp_dir), 'tutorials')
 from pandapower.test.conftest import *
 from pandapower.test.toolbox import *
 from pandapower.test.run_tests import *
-