+++ pandapower/toolbox/elm_selection.py
@@ -8,7 +8,7 @@ import gc
 import numpy as np
 import pandas as pd
 
-from pandapower.toolbox_general_issues import pp_elements, branch_element_bus_dict
+from pandapower.toolbox.general_issues import pp_elements, branch_element_bus_dict
 
 try:
     import pandaplan.core.pplog as logging